Pooled Memory Stream functional

+++ sdk/storage/Azure.Storage.Blobs/tests/BlobClientTests.cs
@@ -820,7 +820,7 @@ namespace Azure.Storage.Blobs.Test
                 MaximumTransferLength = Constants.MB,
                 MaximumConcurrency = 16
             };
-            int size = Constants.Blob.Block.MaxUploadBytes + 1; // ensure that the Parallel upload code path is hit
+            long size = Constants.Blob.Block.MaxUploadBytes + 1; // ensure that the Parallel upload code path is hit
             using var stream = new MemoryStream(GetRandomBuffer(size));
 
             // Act
